requirements.txt

numpy>=1.23
Pillow>=9.5
tqdm>=4.66
# Optional but recommended:
opencv-python>=4.8 ; platform_system != "Linux" or platform_machine != "aarch64"
scikit-image>=0.22
matplotlib>=3.7
scipy>=1.10
scikit-learn>=1.3
